\name{gWidgetsrJava-undocumented}
\alias{gWidgetsrJava-undocumented}


\alias{runHandlerFor}




%% These were copied from gWidgetsrRGtk2
\alias{[-methods}
\alias{[,gWidgetrJava-method}
\alias{[,gCheckboxrJava-method}
\alias{[,gCheckboxgrouprJava-method}
\alias{[,gCommandlinerJava-method}
\alias{[,gDroplistrJava-method}
\alias{[,gEditrJava-method}
\alias{[,GtkTreeView-method}
\alias{[,gGridrJava-method}
\alias{[,gMenurJava-method}
\alias{[,gNotebookrJava-method}
\alias{[,gRadiorJava-method}
\alias{[,rJavaDataFrame-method}
\alias{[,gToolbarrJava-method}
\alias{[,gTreerJava-method}
\alias{[,gVarbrowserrJava-method}
\alias{.delete,gNotebookrJava,guiWidgetsToolkitrJava,ANY-method}
\alias{.delete,gToolbarrJava,guiWidgetsToolkitrJava,ANY-method}

\alias{.svalue<-,gNotebookrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gRadiorJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gSliderrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gSpinbuttonrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gStatusbarrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gTextrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gToolbarrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gAddargrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gBivariaterJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gModelrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gWindowrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gActionrJava,guiWidgetsToolkitrJava,ANY,ANY-method}

%%
%%
%% --  .glabel --
%%
\alias{.glabel-methods}
\alias{.glabel,guiWidgetsToolkitrJava-method} 
%%
%% --  .gbutton --
%%
\alias{.gbutton-methods}
\alias{.gbutton,guiWidgetsToolkitrJava-method} 
\alias{.gbutton,guiWidgetsToolkitrJava,ANY,ANY,ANY,ANY-method} 
\alias{.gbutton,guiWidgetsToolkitrJava,ANY,ANY,ANY,guiWidget-method} 
\alias{.gbutton,guiWidgetsToolkitrJava,ANY,ANY,ANY,guiComponent-method} 
\alias{.gbutton,guiWidgetsToolkitrJava,ANY,ANY,ANY,gActionrJava-method} 
%%
%% --  .gcheckbox --
%%
\alias{.gcheckbox-methods}
\alias{.gcheckbox,guiWidgetsToolkitrJava-method} 
%%
%% --  .gradio --
%%
\alias{.gradio-methods}
\alias{.gradio,guiWidgetsToolkitrJava-method} 
%%
%% --  .gdroplist --
%%
\alias{.gdroplist-methods}
\alias{.gdroplist,guiWidgetsToolkitrJava-method} 
%%
%% --  .gcheckboxgroup --
%%
\alias{.gcheckboxgroup-methods}
\alias{.gcheckboxgroup,guiWidgetsToolkitrJava-method} 
%%
%% --  .gspinbutton --
%%
\alias{.gspinbutton-methods}
\alias{.gspinbutton,guiWidgetsToolkitrJava-method} 
%%
%% --  .gslider --
%%
\alias{.gslider-methods}
\alias{.gslider,guiWidgetsToolkitrJava-method} 
%%
%% --  .gedit --
%%
\alias{.gedit-methods}
\alias{.gedit,guiWidgetsToolkitrJava-method} 
%%
%% --  .gtext --
%%
\alias{.gtext-methods}
\alias{.gtext,guiWidgetsToolkitrJava-method} 
%%
%% --  .gaction --
%%
\alias{.gaction-methods}
\alias{.gaction,guiWidgetsToolkitrJava-method} 
%%
%% --  .gmenu --
%%
\alias{.gmenu-methods}
\alias{.gmenu,guiWidgetsToolkitrJava-method} 
%%
%% --  .gtoolbar --
%%
\alias{.gtoolbar-methods}
\alias{.gtoolbar,guiWidgetsToolkitrJava-method} 
%%
%% --  .gtable --
%%
\alias{.gtable-methods}
\alias{.gtable,guiWidgetsToolkitrJava-method} 
%%
%% --  .gdf --
%%
\alias{.gdf-methods}
\alias{.gdf,guiWidgetsToolkitrJava-method} 
%%
%% --  .gdfnotebook --
%%
\alias{.gdfnotebook-methods}
\alias{.gdfnotebook,guiWidgetsToolkitrJava-method} 
%%
%% --  .gtree --
%%
\alias{.gtree-methods}
\alias{.gtree,guiWidgetsToolkitrJava-method} 
%%
%% --  .gfile --
%%
\alias{.gfile-methods}
\alias{.gfile,guiWidgetsToolkitrJava-method} 
%%
%% --  .gfilebrowse --
%%
\alias{.gfilebrowse-methods}
\alias{.gfilebrowse,guiWidgetsToolkitrJava-method} 
%%
%% --  .gcalendar --
%%
\alias{.gcalendar-methods}
\alias{.gcalendar,guiWidgetsToolkitrJava-method} 
%%
%% --  .ggraphics --
%%
\alias{.ggraphics-methods}
\alias{.ggraphics,guiWidgetsToolkitrJava-method} 
%%
%% --  .ggraphicsnotebook --
%%
\alias{.ggraphicsnotebook-methods}
\alias{.ggraphicsnotebook,guiWidgetsToolkitrJava-method} 
%%
%% --  .gimage --
%%
\alias{.gimage-methods}
\alias{.gimage,guiWidgetsToolkitrJava-method} 
%%
%% --  .gstatusbar --
%%
\alias{.gstatusbar-methods}
\alias{.gstatusbar,guiWidgetsToolkitrJava-method} 
%%
%% --  .gseparator --
%%
\alias{.gseparator-methods}
\alias{.gseparator,guiWidgetsToolkitrJava-method} 
%%
%% --  .gcommandline --
%%
\alias{.gcommandline-methods}
\alias{.gcommandline,guiWidgetsToolkitrJava-method} 
%%
%% --  .ghelp --
%%
\alias{.ghelp-methods}
\alias{.ghelp,guiWidgetsToolkitrJava-method} 
%%
%% --  .ghtml --
%%
\alias{.ghtml-methods}
\alias{.ghtml,guiWidgetsToolkitrJava-method} 
%%
%% --  .ghelpbrowser --
%%
\alias{.ghelpbrowser-methods}
\alias{.ghelpbrowser,guiWidgetsToolkitrJava-method} 
%%
%% --  .ggenericwidget --
%%
\alias{.ggenericwidget-methods}
\alias{.ggenericwidget,guiWidgetsToolkitrJava-method} 
%%
%% --  .gvarbrowser --
%%
\alias{.gvarbrowser-methods}
\alias{.gvarbrowser,guiWidgetsToolkitrJava-method} 
%%
%% --  .gwindow --
%%
\alias{.gwindow-methods}
\alias{.gwindow,guiWidgetsToolkitrJava-method} 
%%
%% --  .ggroup --
%%
\alias{.ggroup-methods}
\alias{.ggroup,guiWidgetsToolkitrJava-method} 
%%
%% --  .gframe --
%%
\alias{.gframe-methods}
\alias{.gframe,guiWidgetsToolkitrJava-method} 
%%
%% --  .gexpandgroup --
%%
\alias{.gexpandgroup-methods}
\alias{.gexpandgroup,guiWidgetsToolkitrJava-method} 
%%
%% --  .gnotebook --
%%
\alias{.gnotebook-methods}
\alias{.gnotebook,guiWidgetsToolkitrJava-method} 
%%
%% --  .glayout --
%%
\alias{.glayout-methods}
\alias{.glayout,guiWidgetsToolkitrJava-method} 
%%
%% --  .gpanedgroup --
%%
\alias{.gpanedgroup-methods}
\alias{.gpanedgroup,guiWidgetsToolkitrJava-method} 
%%
%% --  .addStockIcons --
%%
\alias{.addStockIcons-methods}
\alias{.addStockIcons,guiWidgetsToolkitrJava-method} 
%%
%% --  .getStockIcons --
%%
\alias{.getStockIcons-methods}
\alias{.getStockIcons,guiWidgetsToolkitrJava-method} 
%%
%% -- stockIconFromClass, Object{
\alias{.stockIconFromClass-methods}
\alias{.stockIconFromClass,guiWidgetsToolkitrJava-method}
\alias{.stockIconFromObject-methods}
\alias{.stockIconFromObject,guiWidgetsToolkitrJava-method}

%%
%% --  svalue --
%%
\alias{svalue-methods}
\alias{svalue,guiWidget-method}
\alias{svalue,gWidgetrJava-method}
\alias{svalue,character-method}
\alias{svalue,GtkEntry-method}
\alias{svalue,gSubsetbyrJava-method}
\alias{svalue,GtkTreeViewColumn-method} 
%%
%% --  .svalue --
%%
\alias{.svalue-methods}
\alias{.svalue,character,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gActionrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gButtonrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gCheckboxrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gCheckboxgrouprJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gCommandlinerJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gDroplistrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gEditrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,GtkEntry,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gExpandgrouprJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,GtkTreeView,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gGridrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gGenericWidgetrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gGenericWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gHelprJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gImagerJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gLabelrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gMenurJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gNotebookrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gPanedgrouprJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gRadiorJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gSliderrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gSpinbuttonrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gStatusbarrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gTextrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gToolbarrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gTreerJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gVarbrowserrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gAddargrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gUnivariaterJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gUnivariateTablerJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gFileURLrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gBivariaterJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gModelrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gLmerrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gWindowrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gCalendarrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,gTablerJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.svalue,character,guiWidgetsToolkitrJava-method}
\alias{.svalue,gAddargrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gBivariaterJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gButtonrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gCalendarrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gCheckboxgrouprJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gCheckboxrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gCommandlinerJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gDroplistrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gEditrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gExpandgrouprJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gFileURLrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gHelprJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gHtmlrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gImagerJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gLabelrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gLmerrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gMenurJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gModelrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gNotebookrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gPanedgrouprJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gRadiorJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gSliderrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gSpinbuttonrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gStatusbarrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gTablerJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gTableWithFilterrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gTextrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gToolbarrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gTreerJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gUnivariaterJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gUnivariateTablerJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gVarbrowserrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gWindowrJava,guiWidgetsToolkitrJava-method}


%%
%% --  svalue<- --
%%
\alias{svalue<--methods}
\alias{svalue<-,guiWidget-method}
\alias{svalue<-,gWidgetrJava-method} 
%%
%% --  .svalue<- --
%%
\alias{.svalue<--methods}
\alias{.svalue<-,gButtonrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gCheckboxrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gCheckboxgrouprJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gCommandlinerJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gDroplistrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gEditrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,GtkEntry,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gExpandgrouprJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gGraphicsrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gGridrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gGrouprJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gImagerJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gLabelrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gMenurJava,guiWidgetsToolkitrJava,ANY,list-method}
\alias{.svalue<-,gMenurJava,guiWidgetsToolkitrJava,ANY,gMenurJava-method}
\alias{.svalue<-,gMenurJava,guiWidgetsToolkitrJava,ANY,guiWidget-method}
\alias{.svalue<-,gPanedgrouprJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gNotebookrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue<-,gRadiorJava,guiWidgetsToolkitrJava-method}
\alias{.svalue<-,gSliderrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue<-,gSpinbuttonrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue<-,gStatusbarrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue<-,gTextrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue<-,gToolbarrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue<-,gAddargrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue<-,gBivariaterJava,guiWidgetsToolkitrJava-method}
\alias{.svalue<-,gModelrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue<-,gWindowrJava,guiWidgetsToolkitrJava-method} 
\alias{.svalue<-,gFramerJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue<-,gHtmlrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
%%
%% --  .leftBracket --
%%
\alias{.leftBracket-methods}
\alias{.leftBracket,gCheckboxrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gCheckboxgrouprJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gCommandlinerJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gDroplistrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gEditrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,GtkTreeView,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gMenurJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gNotebookrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gRadiorJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,rJavaDataFrame,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gToolbarrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gTreerJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gVarbrowserrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  .leftBracket<- --
%%
\alias{.leftBracket<--methods}
\alias{.leftBracket<-,gCheckboxrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket<-,gCheckboxgrouprJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket<-,gDroplistrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket<-,gEditrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket<-,GtkTreeView,guiWidgetsToolkitrJava-method}
\alias{.leftBracket<-,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket<-,gLayoutrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket<-,gMenurJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket<-,gNotebookrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket<-,gRadiorJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket<-,gToolbarrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  add --
%%
\alias{add-methods}
\alias{add,guiWidget-method}
\alias{add,gWidgetrJava-method} 
%%
%% --  .add --
%%
\alias{.add-methods}
\alias{.add,guiWidget,guiWidgetsToolkitrJava,ANY-method}
\alias{.add,guiWidget,guiWidgetsToolkitrJava,guiWidgetORgWidgetrJavaORrJavaObject-method}
\alias{.add,gBasicDialogNoParentrJava,guiWidgetsToolkitrJava,guiWidget-method}
\alias{.add,gBasicDialogNoParentrJava,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.add,gContainerrJava,guiWidgetsToolkitrJava,guiWidget-method}
\alias{.add,gContainerrJava,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.add,gContainerrJava,guiWidgetsToolkitrJava,gGraphicsrJava-method}
\alias{.add,gContainerrJava,guiWidgetsToolkitrJava,gMenurJava-method}
\alias{.add,gDfNotebookrJava,guiWidgetsToolkitrJava,ANY-method}
\alias{.add,gExpandgrouprJava,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.add,gGrouprJava,guiWidgetsToolkitrJava,gGraphicsrJava-method}
\alias{.add,gGrouprJava,guiWidgetsToolkitrJava,rJavaObject-method}
\alias{.add,gHelprJava,guiWidgetsToolkitrJava,ANY-method}
\alias{.add,gHelpbrowserrJava,guiWidgetsToolkitrJava,ANY-method}
\alias{.add,gMenurJava,guiWidgetsToolkitrJava,guiWidget-method}
\alias{.add,gMenurJava,guiWidgetsToolkitrJava,gMenurJava-method}
\alias{.add,gMenurJava,guiWidgetsToolkitrJava,list-method}
\alias{.add,gLayoutrJava,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.add,gAddrJava,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.add,gNotebookrJava,guiWidgetsToolkitrJava,guiWidget-method}
\alias{.add,gNotebookrJava,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.add,gTextrJava,guiWidgetsToolkitrJava,character-method}
\alias{.add,gTextrJava,guiWidgetsToolkitrJava,guiWidget-method}
\alias{.add,gTextrJava,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.add,gToolbarrJava,guiWidgetsToolkitrJava,list-method}
\alias{.add,gWindowrJava,guiWidgetsToolkitrJava,rJavaObject-method}
\alias{.add,gWindowrJava,guiWidgetsToolkitrJava,gMenurJava-method}
\alias{.add,gWindowrJava,guiWidgetsToolkitrJava,gToolbarrJava-method}
\alias{.add,gWindowrJava,guiWidgetsToolkitrJava,gStatusbarrJava-method}
\alias{.add,gWindowrJava,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.add,gPanedgrouprJava,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.add,gWidgetrJava,guiWidgetsToolkitrJava,try-error-method} 
%%
%% --  addSpace --
%%
\alias{addSpace-methods}
\alias{addSpace,guiWidget-method}
\alias{addSpace,gWidgetrJava-method} 
%%
%% --  .addSpace --
%%
\alias{.addSpace-methods}
\alias{.addSpace,gContainerrJava,guiWidgetsToolkitrJava-method}
\alias{.addSpace,gExpandgrouprJava,guiWidgetsToolkitrJava-method}
%%
%% --  addSpring --
%%
\alias{addSpring-methods}
\alias{addSpring,guiWidget-method}
\alias{addSpring,gWidgetrJava-method} 
%%
%% --  .addSpring --
%%
\alias{.addSpring-methods}
\alias{.addSpring,gContainerrJava,guiWidgetsToolkitrJava-method} 
\alias{.addSpring,gExpandgrouprJava,guiWidgetsToolkitrJava-method}
%%
%% --  delete --
%%
\alias{delete-methods}
\alias{delete,guiWidget-method}
\alias{delete,gWidgetrJava-method} 
%%
%% --  .delete --
%%
\alias{.delete-methods}
\alias{.delete,gContainerrJava,guiWidgetsToolkitrJava,guiWidget-method}
\alias{.delete,gContainerrJava,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.delete,rJavaObject,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.delete,rJavaObject,guiWidgetsToolkitrJava,guiWidget-method}
\alias{.delete,rJavaObject,guiWidgetsToolkitrJava,rJavaObject-method}
\alias{.delete,gWidgetrJava,guiWidgetsToolkitrJava,rJavaObject-method}
\alias{.delete,gMenurJava,guiWidgetsToolkitrJava,guiWidget-method}
\alias{.delete,gMenurJava,guiWidgetsToolkitrJava,gWidgetrJava-method}
\alias{.delete,gMenurJava,guiWidgetsToolkitrJava,gMenurJava-method}
\alias{.delete,gMenurJava,guiWidgetsToolkitrJava,list-method}
\alias{.delete,gNotebookrJava,guiWidgetsToolkitrJava-method}
\alias{.delete,gToolbarrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  dispose --
%%
\alias{dispose-methods}
\alias{dispose,guiWidget-method}
\alias{dispose,gWidgetrJava-method}
\alias{dispose,gTextrJava-method} 
%%
%% --  .dispose --
%%
\alias{.dispose-methods}
\alias{.dispose,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.dispose,gNotebookrJava,guiWidgetsToolkitrJava-method}
\alias{.dispose,gTextrJava,guiWidgetsToolkitrJava-method}
\alias{.dispose,gHelprJava,guiWidgetsToolkitrJava-method}
\alias{.dispose,gWindowrJava,guiWidgetsToolkitrJava-method}

%%
%% --  visible --
%%
\alias{visible-methods}
\alias{visible,guiWidget-method}
\alias{visible,gWidgetrJava-method} 
%%
%% --  .visible --
%%
\alias{.visible-methods}
\alias{.visible,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.visible,gBasicDialogNoParentrJava,guiWidgetsToolkitrJava-method}
\alias{.visible,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.visible,gExpandgrouprJava,guiWidgetsToolkitrJava-method}
\alias{.visible,gLayoutrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  visible<- --
%%
\alias{visible<--methods}
\alias{visible<-,guiWidget-method}
\alias{visible<-,gWidgetrJava-method} 
%%
%% --  .visible<- --
%%
\alias{.visible<--methods}
\alias{.visible<-,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.visible<-,gExpandgrouprJava,guiWidgetsToolkitrJava-method}
\alias{.visible<-,gGraphicsrJava,guiWidgetsToolkitrJava-method}
\alias{.visible<-,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.visible<-,gLayoutrJava,guiWidgetsToolkitrJava-method}
\alias{.visible<-,gWindowrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  enabled --
%%
\alias{enabled-methods}
\alias{enabled,guiWidget-method}
\alias{enabled,gWidgetrJava-method} 
%%
%% --  .enabled --
%%
\alias{.enabled-methods}
\alias{.enabled,gWidgetrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  enabled<- --
%%
\alias{enabled<--methods}
\alias{enabled<-,guiWidget-method}
\alias{enabled<-,gWidgetrJava-method} 
%%
%% --  .enabled<- --
%%
\alias{.enabled<--methods}
\alias{.enabled<-,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.enabled<-,rJavaObject,guiWidgetsToolkitrJava-method} 
\alias{.enabled<-,gActionrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  .insert- --
%%
\alias{.insert-methods}
\alias{.insert,gTextrJava,guiWidgetsToolkitrJava-method}
%%
%% --  size --
%%
\alias{size-methods}
\alias{size,guiWidget-method}
\alias{size,gWidgetrJava-method} 
%%
%% --  .size --
%%
\alias{.size-methods}
\alias{.size,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.size,gWindowrJava,guiWidgetsToolkitrJava-method}
%%
%% --  size<- --
%%
\alias{size<--methods}
\alias{size<-,guiWidget-method}
\alias{size<-,gWidgetrJava-method} 
%%
%% --  .size<- --
%%
\alias{.size<--methods}
\alias{.size<-,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.size<-,gNotebookrJava,guiWidgetsToolkitrJava-method}
\alias{.size<-,gGridrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  focus --
%%
\alias{focus-methods}
\alias{focus,guiWidget-method}
\alias{focus,gWidgetrJava-method} 
%%
%% --  .focus --
%%
\alias{.focus-methods}
\alias{.focus,gWidgetrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  focus<- --
%%
\alias{focus<--methods}
\alias{focus<-,guiWidget-method}
\alias{focus<-,gWidgetrJava-method} 
%%
%% --  .focus<- --
%%
\alias{.focus<--methods}
\alias{.focus<-,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.focus<-,rJavaObject,guiWidgetsToolkitrJava-method} 
%%
%% --  tooltip<- --
%%
\alias{tooltip<--methods}
\alias{tooltip<-,guiWidget-method}
\alias{tooltip<-,gWidgetrJava-method} 
%%
%% --  .tooltip<- --
%%
\alias{.tooltip<--methods}
\alias{.tooltip<-,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.tooltip<-,rJavaObject,guiWidgetsToolkitrJava-method} 

%%
%% --  defaultWidget --
%%
\alias{defaultWidget-methods}
\alias{defaultWidget,guiWidget-method}
\alias{defaultWidget,gWidgetrJava-method} 
%%
%% --  .defaultWidget --
%%
\alias{.defaultWidget-methods}
\alias{.defaultWidget,gWidgetrJava,guiWidgetsToolkitrJava-method}

%%
%% --  defaultWidget<- --
%%
\alias{defaultWidget<--methods}
\alias{defaultWidget<-,guiWidget-method}
\alias{defaultWidget<-,gWidgetrJava-method} 
%%
%% --  .defaultWidget<- --
%%
\alias{.defaultWidget<--methods}
\alias{.defaultWidget<-,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.defaultWidget<-,rJavaObject,guiWidgetsToolkitrJava-method} 

%%
%% --  font --
%%
\alias{font-methods}
\alias{font,guiWidget-method}
\alias{font,gWidgetrJava-method} 
%%
%% --  .font --
%%
\alias{.font-methods}
\alias{.font,GtkWindow,guiWidgetsToolkitrJava-method} 
%%
%% --  font<- --
%%
\alias{font<--methods}
\alias{font<-,guiWidget-method}
\alias{font<-,gWidgetrJava-method} 
%%
%% --  .font<- --
%%
\alias{.font<--methods}
\alias{.font<-,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.font<-,rJavaObject,guiWidgetsToolkitrJava-method}
\alias{.font<-,gTextrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  tag --
%%
\alias{tag-methods}
\alias{tag,guiWidget-method}
\alias{tag,gWidgetrJava-method}
\alias{tag,rJavaObject-method} 
%%
%% --  .tag --
%%
\alias{.tag-methods}
\alias{.tag,guiWidget,guiWidgetsToolkitrJava-method}
\alias{.tag,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.tag,rJavaObject,guiWidgetsToolkitrJava-method} 
%%
%% --  tag<- --
%%
\alias{tag<--methods}
\alias{tag<-,guiWidget-method}
\alias{tag<-,gWidgetrJava-method}
\alias{tag<-,rJavaObject-method} 
%%
%% --  .tag<- --
%%
\alias{.tag<--methods}
\alias{.tag<-,guiWidget,guiWidgetsToolkitrJava-method}
\alias{.tag<-,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.tag<-,rJavaObject,guiWidgetsToolkitrJava-method} 
%%
%% --  id --
%%
\alias{id-methods}
\alias{id,guiWidget-method}
\alias{id,gWidgetrJava-method}
\alias{id,rJavaObject-method}
\alias{id,ANY-method}
\alias{id,GtkTreeViewColumn-method} 
%%
%% --  .id --
%%
\alias{.id-methods}
\alias{.id,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.id,rJavaObject,guiWidgetsToolkitrJava-method} 
%%
%% --  id<- --
%%
\alias{id<--methods}
\alias{id<-,guiWidget-method}
\alias{id<-,gWidgetrJava-method}
\alias{id<-,rJavaObject-method}
\alias{id<-,ANY-method}
\alias{id<-,GtkTreeViewColumn-method} 
%%
%% --  .id<- --
%%
\alias{.id<--methods}
\alias{.id<-,gWidgetrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  removehandler --
%%
\alias{removehandler-methods}
\alias{removehandler,guiWidget-method}
\alias{removehandler,gWidgetrJava-method}
\alias{removehandler,rJavaObject-method}
\alias{removehandler,GtkTreeViewColumn-method} 
%%
%% --  .removehandler --
%%
\alias{.removehandler-methods}
\alias{.removehandler,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.removehandler,rJavaObject,guiWidgetsToolkitrJava-method} 

%%
%% --  addhandler --
%%
\alias{addhandler-methods}
\alias{addhandler,guiWidget-method}
\alias{addhandler,gWidgetrJava-method}
\alias{addhandler,rJavaObject-method}

%% --  blockhandler --
%%
\alias{blockhandler-methods}
\alias{blockhandler,gWidgetrJava-method}
\alias{blockhandler,rJavaObject-method}
%%
%% --  .blockhandler --
%%
\alias{.blockhandler-methods}
\alias{.blockhandler,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.blockhandler,rJavaObject,guiWidgetsToolkitrJava-method} 

%% --  unblockhandler --
%%
\alias{unblockhandler-methods}
\alias{unblockhandler,gWidgetrJava-method}
\alias{unblockhandler,rJavaObject-method}
%%
%% --  .unblockhandler --
%%
\alias{.unblockhandler-methods}
\alias{.unblockhandler,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.unblockhandler,rJavaObject,guiWidgetsToolkitrJava-method} 

%%
%% --  addhandlerchanged --
%%
\alias{addhandlerchanged-methods}
\alias{addhandlerchanged,guiWidget-method}
\alias{addhandlerchanged,gWidgetrJava-method}
\alias{addhandlerchanged,rJavaObject-method}
\alias{addhandlerchanged,ANY-method}
\alias{addhandlerchanged,gSubsetbyrJava-method}
%%
%% --  .addhandlerchanged --
%%
\alias{.addhandlerchanged-methods}
\alias{.addhandlerchanged,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gCheckboxrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gEditrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gExpandgrouprJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gSubsetbyrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gLabelrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gNotebookrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gRadiorJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gSliderrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gSpinbuttonrJava,guiWidgetsToolkitrJava-method} 
\alias{.addhandlerchanged,gButtonrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gCalendarrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gCheckboxgrouprJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gDroplistrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gImagerJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerchanged,gTablerJava,guiWidgetsToolkitrJava-method}
%%
%% --  addhandlerkeystroke --
%%
\alias{addhandlerkeystroke-methods}
\alias{addhandlerkeystroke,guiWidget-method}
\alias{addhandlerkeystroke,gWidgetrJava-method}
\alias{addhandlerkeystroke,rJavaObject-method} 
%%
%% --  .addhandlerkeystroke --
%%
\alias{.addhandlerkeystroke-methods}
\alias{.addhandlerkeystroke,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerkeystroke,gEditrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerkeystroke,gTextrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  addhandlerclicked --
%%
\alias{addhandlerclicked-methods}
\alias{addhandlerclicked,guiWidget-method}
\alias{addhandlerclicked,gWidgetrJava-method}
\alias{addhandlerclicked,rJavaObject-method} 
%%
%% --  .addhandlerclicked --
%%
\alias{.addhandlerclicked-methods}
\alias{.addhandlerclicked,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerclicked,gButtonrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerclicked,gDroplistrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerclicked,gGraphicsrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerclicked,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerclicked,gImagerJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerclicked,gLabelrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerclicked,gTablerJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerclicked,gCheckboxrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerclicked,gRadiorJava,guiWidgetsToolkitrJava-method}
%%
%% --  addhandlerdoubleclick --
%%
\alias{addhandlerdoubleclick-methods}
\alias{addhandlerdoubleclick,guiWidget-method}
\alias{addhandlerdoubleclick,gWidgetrJava-method}
\alias{addhandlerdoubleclick,rJavaObject-method} 
%%
%% --  .addhandlerdoubleclick --
%%
\alias{.addhandlerdoubleclick-methods}
\alias{.addhandlerdoubleclick,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerdoubleclick,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerdoubleclick,gTreerJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerdoubleclick,gTablerJava,guiWidgetsToolkitrJava-method} 
%%
%% --  addhandlerrightclick --
%%
\alias{addhandlerrightclick-methods}
\alias{addhandlerrightclick,guiWidget-method}
\alias{addhandlerrightclick,gWidgetrJava-method}
\alias{addhandlerrightclick,rJavaObject-method} 
%%
%% --  .addhandlerrightclick --
%%
\alias{.addhandlerrightclick-methods}
\alias{.addhandlerrightclick,gWidgetrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  addhandlerdestroy --
%%
\alias{addhandlerdestroy-methods}
\alias{addhandlerdestroy,guiWidget-method}
\alias{addhandlerdestroy,gWidgetrJava-method}
\alias{addhandlerdestroy,rJavaObject-method} 
%%
%% --  .addhandlerdestroy --
%%
\alias{.addhandlerdestroy-methods}
\alias{.addhandlerdestroy,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerdestroy,gWindowrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  addhandlerexpose --
%%
\alias{addhandlerexpose-methods}
\alias{addhandlerexpose,guiWidget-method}
\alias{addhandlerexpose,gWidgetrJava-method}
\alias{addhandlerexpose,rJavaObject-method} 
%%
%% --  .addhandlerexpose --
%%
\alias{.addhandlerexpose-methods}
\alias{.addhandlerexpose,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerexpose,gComponentrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerexpose,gGraphicsrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerexpose,gNotebookrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  addhandlerunrealize --
%%
\alias{addhandlerunrealize-methods}
\alias{addhandlerunrealize,guiWidget-method}
\alias{addhandlerunrealize,gWidgetrJava-method}
\alias{addhandlerunrealize,rJavaObject-method} 
%%
%% --  .addhandlerunrealize --
%%
\alias{.addhandlerunrealize-methods}
\alias{.addhandlerunrealize,gWidgetrJava,guiWidgetsToolkitrJava-method} 

%%
%% --  addhandleridle --
%%
\alias{addhandleridle-methods}
\alias{addhandleridle,guiWidget-method}
\alias{addhandleridle,gWidgetrJava-method}
\alias{addhandleridle,rJavaObject-method} 
%%
%% --  .addhandleridle --
%%
\alias{.addhandleridle-methods}
\alias{.addhandleridle,gWidgetrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  addpopupmenu --
%%
\alias{addpopupmenu-methods}
\alias{addpopupmenu,guiWidget-method}
\alias{addpopupmenu,gWidgetrJava-method}
\alias{addpopupmenu,rJavaObject-method}
%%
%% --  .addpopupmenu --
%%
\alias{.addpopupmenu-methods}
\alias{.addpopupmenu,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.addpopupmenu,gButtonrJava,guiWidgetsToolkitrJava-method} 
\alias{.addpopupmenu,gLabelrJava,guiWidgetsToolkitrJava-method}
%%
%% --  add3rdmousepopupmenu --
%%
\alias{add3rdmousepopupmenu-methods}
\alias{add3rdmousepopupmenu,guiWidget-method}
\alias{add3rdmousepopupmenu,gWidgetrJava-method}
\alias{add3rdmousepopupmenu,rJavaObject-method} 
%%
%% --  .add3rdmousepopupmenu --
%%
\alias{.add3rdmousepopupmenu-methods}
\alias{.add3rdmousepopupmenu,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.add3rdmousepopupmenu,rJavaObject,guiWidgetsToolkitrJava-method} 
\alias{.add3rdmousepopupmenu,gLabelrJava,guiWidgetsToolkitrJava-method}

%%
%% --  addhandlermousemotion --
%%
\alias{addhandlermousemotion-methods}
\alias{addhandlermousemotion,gWidgetrJava-method}
\alias{addhandlermousemotion,rJavaObject-method} 
%%
%% --  .addhandlermousemotion --
%%
\alias{.addhandlermousemotion-methods}
\alias{.addhandlermousemotion,gWidgetrJava,guiWidgetsToolkitrJava-method} 

%%
%% --  addhandlerblur --
%%
\alias{addhandlerblur-methods}
\alias{addhandlerblur,gWidgetrJava-method}
\alias{addhandlerblur,rJavaObject-method} 
%%
%% --  .addhandlerblur --
%%
\alias{.addhandlerblur-methods}
\alias{.addhandlerblur,gWidgetrJava,guiWidgetsToolkitrJava-method} 

%%
%% --  adddropsource --
%%
\alias{adddropsource-methods}
\alias{adddropsource,guiWidget-method}
\alias{adddropsource,gWidgetrJava-method}
\alias{adddropsource,rJavaObject-method} 
%%
%% --  .adddropsource --
%%
\alias{.adddropsource-methods}
\alias{.adddropsource,rJavaObject,guiWidgetsToolkitrJava-method}
\alias{.adddropsource,gWidgetrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  adddropmotion --
%%
\alias{adddropmotion-methods}
\alias{adddropmotion,guiWidget-method}
\alias{adddropmotion,gWidgetrJava-method}
\alias{adddropmotion,rJavaObject-method} 
%%
%% --  .adddropmotion --
%%
\alias{.adddropmotion-methods}
\alias{.adddropmotion,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.adddropmotion,rJavaObject,guiWidgetsToolkitrJava-method} 
%%
%% --  adddroptarget --
%%
\alias{adddroptarget-methods}
\alias{adddroptarget,guiWidget-method}
\alias{adddroptarget,gWidgetrJava-method}
\alias{adddroptarget,rJavaObject-method}
\alias{adddroptarget,gImagerJava,guiWidgetsToolkitrJava-method}
%%
%% --  .adddroptarget --
%%
\alias{.adddroptarget-methods}
\alias{.adddroptarget,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.adddroptarget,rJavaObject,guiWidgetsToolkitrJava-method}
\alias{.adddroptarget,gLabelrJava,guiWidgetsToolkitrJava-method}
\alias{.adddroptarget,gImagerJava,guiWidgetsToolkitrJava-method} 
%%
%% --  .galert --
%%
\alias{.galert-methods}
\alias{.galert,guiWidgetsToolkitrJava-method} 
%%
%% --  .gmessage --
%%
\alias{.gmessage-methods}
\alias{.gmessage,guiWidgetsToolkitrJava-method} 
%%
%% --  .ginput --
%%
\alias{.ginput-methods}
\alias{.ginput,guiWidgetsToolkitrJava-method} 
%%
%% --  .gconfirm --
%%
\alias{.gconfirm-methods}
\alias{.gconfirm,guiWidgetsToolkitrJava-method} 
%%
%% --  .gbasicdialog --
%%
\alias{.gbasicdialog-methods}
\alias{.gbasicdialog,guiWidgetsToolkitrJava-method} 
%%
%% --  .gbasicdialognoparent --
%%
\alias{.gbasicdialognoparent-methods}
\alias{.gbasicdialognoparent,guiWidgetsToolkitrJava-method} 
%%
%% --  update --
%%
\alias{update-methods}
\alias{update,ANY-method}
\alias{update,guiWidget-method}
\alias{update,gWidgetrJava-method}
\alias{update,gSubsetbyrJava-method}
\alias{update,gTreerJava-method} 
%%
%% --  .update --
%%
\alias{.update-methods}
\alias{.update,gComponentrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  .length --
%%
\alias{.length-methods}
\alias{.length,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.length,gCheckboxgrouprJava,guiWidgetsToolkitrJava-method}
\alias{.length,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.length,gNotebookrJava,guiWidgetsToolkitrJava-method}
\alias{.length,gDroplistrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  dim --
%%
\alias{dim-methods}
\alias{dim,ANY-method}
\alias{dim,guiWidget-method}
\alias{dim,gWidgetrJava-method} 
%%
%% --  .dim --
%%
\alias{.dim-methods}
\alias{.dim,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.dim,gWidgetrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  .dimnames --
%%
\alias{.dimnames-methods}
\alias{.dimnames,gGridrJava,guiWidgetsToolkitrJava-method}
%%
%% --  .dimnames<- --
%%
\alias{.dimnames<--methods}
\alias{.dimnames<-,gGridrJava,guiWidgetsToolkitrJava-method} 
\alias{.dimnames<-,gDfrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  names --
%%
\alias{names-methods}
\alias{names,ANY-method}
\alias{names,guiWidget-method}
\alias{names,gWidgetrJava-method} 
%%
%% --  .names --
%%
\alias{.names-methods}
\alias{.names,gCheckboxgrouprJava,guiWidgetsToolkitrJava-method}
\alias{.names,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.names,gFramerJava,guiWidgetsToolkitrJava-method}
\alias{.names,gNotebookrJava,guiWidgetsToolkitrJava-method} 
%%
%% --  names<- --
%%
\alias{names<--methods}
\alias{names<-,ANY-method}
\alias{names<-,guiWidget-method}
\alias{names<-,gWidgetrJava-method} 
%%
%% --  .names<- --
%%
\alias{.names<--methods}
\alias{.names<-,gGridrJava,guiWidgetsToolkitrJava-method}
\alias{.names<-,gFramerJava,guiWidgetsToolkitrJava-method} 
\alias{.names<-,gNotebookrJava,guiWidgetsToolkitrJava-method} 
%%
%%
%% -- .getToolkitWidget --
\alias{.getToolkitWidget,gWidgetrJava,guiWidgetsToolkitrJava-method}
\alias{.getToolkitWidget,gActionrJava,guiWidgetsToolkitrJava-method}
%%
%% -- [<-
\alias{[<-,gWidgetrJava-method}
\alias{[<-,gCheckboxrJava-method}
\alias{[<-,gCheckboxgrouprJava-method}
\alias{[<-,gDroplistrJava-method}
\alias{[<-,gEditrJava-method}
\alias{[<-,GtkTreeView-method}
\alias{[<-,gGridrJava-method}
\alias{[<-,gLayoutrJava-method}
\alias{[<-,gMenurJava-method}
\alias{[<-,gNotebookrJava-method}
\alias{[<-,gRadiorJava-method}
\alias{[<-,gToolbarrJava-method}
%% -- dimnames
\alias{dimnames,gWidgetrJava-method}
\alias{dimnames<-,gWidgetrJava-method}
\alias{length,gWidgetrJava-method}
\alias{length,gSubsetbyrJava-method}
\alias{length,gDroplistrJava-method}

%% gtablwithfilter
\alias{.addhandlerchanged,gTableWithFilterrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerclicked,gTableWithFilterrJava,guiWidgetsToolkitrJava-method}
\alias{.addhandlerdoubleclick,gTableWithFilterrJava,guiWidgetsToolkitrJava-method}
\alias{[<-,gTableWithFilterrJava-method}
\alias{[,gTableWithFilterrJava-method}
\alias{.dim,gTableWithFilterrJava,guiWidgetsToolkitrJava'-method}
\alias{.ggenericwidget,ANY}
\alias{.leftBracke,gTableWithFilterrJava,guiWidgetsToolkitrJava-method}
\alias{.leftBracket,gTableWithFilterrJava,guiWidgetsToolkitrJava-method}
\alias{.length,gRadiorJava,guiWidgetsToolkitrJava}
\alias{.length,gTableWithFilterrJava,guiWidgetsToolkitrJava-method}
\alias{.siz,gTableWithFilterrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue,gTableWithFilterrJava,guiWidgetsToolkitrJava,ANY,ANY-method}
\alias{.svalue,gTableWithFilterrJava,ANY,ANY,guiWidgetsToolkitrJava-method}
\alias{.dim,gTableWithFilterrJava,guiWidgetsToolkitrJava-method}
\alias{.ggenericwidget,ANY-method}
\alias{.leftBracket<-,gTableWithFilterrJava,guiWidgetsToolkitrJava-method}
\alias{.length,gRadiorJava,guiWidgetsToolkitrJava-method}
\alias{.size<-,gTableWithFilterrJava,guiWidgetsToolkitrJava-method}
\alias{.svalue<-,gTableWithFilterrJava,guiWidgetsToolkitrJava,ANY,ANY-method}


\title{Undocumented functions in gWidgetsrJava}

\description{

	The gWidgetsrJava package implements the gWidgets API. In
	gWidgets the primary constructors and methods are
	documented. The \pkg{gWidgets} package defines the generics,
	and the dispatch is handled by methods defined in the toolkit
	implementations. There are two levels of dispatch. The method
	\code{method} dispatches to the generic \code{.method} based
	on the toolkit, and this dispatches within the toolkit. As the
	toolkits simply provide the objects \code{.method}, there is
	no reason to document them in the toolkits.

}
\keyword{interface}% at least one, from doc/KEYWORDS
